requirements.txt
psycopg2-binary==2.9.7
pandas==2.0.3
dash==2.11.1
python-dotenv==1.0.0
